Change update propagation to reuse the revalidate node functionality rather than a separate function
