+++ datajunction-server/datajunction_server/database/history.py
@@ -24,6 +24,7 @@ class ActivityType(StrEnum):
     TAG = "tag"
     SET_ATTRIBUTE = "set_attribute"
     STATUS_CHANGE = "status_change"
+    UPSTREAM_UPDATE = "upstream_update"
 
 
 class EntityType(StrEnum):
